+++ fftconvolver.h
@@ -14,8 +14,8 @@ class FFTConvolver {
    * sizes. In the case of even image sizes, the middle falls between two
    * pixels.
    */
-  static void Convolve(class FFTWManager& fftw, double* image, size_t imgWidth,
-                       size_t imgHeight, const double* kernel,
+  static void Convolve(class FFTWManager& fftw, float* image, size_t imgWidth,
+                       size_t imgHeight, const float* kernel,
                        size_t kernelSize);
 
   /**
@@ -25,9 +25,9 @@ class FFTConvolver {
    * This function assumes that n/2 is the middle pixel, and reverses the
    * kernel.
    */
-  static void ReverseAndConvolve(class FFTWManager& fftw, double* image,
+  static void ReverseAndConvolve(class FFTWManager& fftw, float* image,
                                  size_t imgWidth, size_t imgHeight,
-                                 const double* kernel, size_t kernelSize);
+                                 const float* kernel, size_t kernelSize);
 
   /**
    * Prepare a smaller kernel for convolution with ConvolveSameSize. When the
@@ -35,25 +35,24 @@ class FFTConvolver {
    * once and multiple times ConvolveSameSize(), than calling Convolve()
    * multiple times.
    */
-  static void PrepareSmallKernel(double* dest, size_t imgWidth,
-                                 size_t imgHeight, const double* kernel,
-                                 size_t kernelSize);
+  static void PrepareSmallKernel(float* dest, size_t imgWidth, size_t imgHeight,
+                                 const float* kernel, size_t kernelSize);
   /**
    * Prepare a kernel for convolution with ConvolveSameSize(). The kernel should
    * be already of the same size as the image to be convolved, or otherwise
    * PrepareSmallKernel() should be used.
    */
-  static void PrepareKernel(double* dest, const double* source, size_t imgWidth,
+  static void PrepareKernel(float* dest, const float* source, size_t imgWidth,
                             size_t imgHeight);
 
   /**
    * Convolve an image with an already prepared kernel of the same size.
    */
-  static void ConvolveSameSize(class FFTWManager& fftw, double* image,
-                               const double* kernel, size_t imgWidth,
+  static void ConvolveSameSize(class FFTWManager& fftw, float* image,
+                               const float* kernel, size_t imgWidth,
                                size_t imgHeight);
 
-  static void Reverse(double* image, size_t imgWidth, size_t imgHeight);
+  static void Reverse(float* image, size_t imgWidth, size_t imgHeight);
 
  private:
 };
